requirements.txt
torch
tqdm
psutil
matplotlib
tensorboard
wandb
dataclasses
pillow
torchvision
moviepy
plotly
python-hostlist
scipy
scikit-learn
sentencepiece
